# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Restaurant Management API
  description: |
    Restaurant Management API for menu item, user type, restaurants management.
    Phase 2 contract-first specification.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/v1/users:
    get:
      tags: [ Users ]
      operationId: listUsers
      summary: List active users
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Optional name filter
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Users returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [ Users ]
      operationId: createUser
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email or login already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/v1/users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [ Users ]
      operationId: getUserById
      summary: Get user details
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [ Users ]
      operationId: updateUser
      summary: Update user profile
      description: Updates non-password user information only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [ Users ]
      operationId: deleteUser
      summary: Delete user account
      responses:
        '204':
          description: User deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/user-types:
    get:
      summary: List user types
      responses:
        '200':
          description: List returned successfully
    post:
      summary: Create a user type
      responses:
        '201':
          description: User type created
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/user-types/{id}:
    get:
      summary: Get a user type by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User type found
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a user type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User type updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a user type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User type deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/restaurants:
    get:
      summary: List restaurants
      responses:
        '200':
          description: List returned successfully
    post:
      summary: Create a restaurant
      responses:
        '201':
          description: Restaurant created
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/restaurants/{id}:
    get:
      summary: Get a restaurant by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restaurant found
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restaurant updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Restaurant deleted
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/restaurants/{restaurantId}/menu-items:
    get:
      summary: List menu items for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List returned successfully
    post:
      summary: Create a menu item for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Menu item created
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/menu-items/{id}:
    get:
      summary: Get a menu item by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Menu item found
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Menu item updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Menu item deleted
components:
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

  schemas:
    UserResponse:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        login:
          type: string
          example: john.doe
        type:
          $ref: '#/components/schemas/UserType'
        createdAt:
          type: string
          format: date-time
        lastModifiedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        login:
          type: string
          example: john.doe
        type:
          $ref: '#/components/schemas/UserType'
      required: [ name, email, login, type ]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
    Address:
      type: object
      properties:
        street:
          type: string
          example: Main Street
        number:
          type: string
          example: '150'
        city:
          type: string
          example: Springfield
        zipCode:
          type: string
          example: '62704'
      required: [ street, number, city, zipCode ]
    UserType:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: OWNER
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string